created separate mongodb files for connection, validation and other method - removed wrong indexex work - added test to validation
